backend/requirements.txt
alembic==1.9.4
async-timeout==4.0.2
bcrypt==4.0.1
cachelib==0.10.2
click==8.1.3
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mysql==0.0.3
mysqlclient==2.1.1
python-dotenv==0.21.1
redis==4.5.1
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.14.0
